requirements.txt
Django==1.10.3
djangorestframework==3.5.3
django-rest-swagger==2.1.0
django-rest-framework-docs==0.1.7